Refactor sidepanel directory; fix jest tests
