+++ src/javascript/ContentEditor/registerActions.jsx
@@ -3,7 +3,7 @@ import {ArrowLeft} from '@jahia/moonstone';
 import {registerEditActions} from '~/ContentEditor/actions/registerEditActions';
 import {registerCreateActions} from '~/ContentEditor/actions/registerCreateActions';
 import {registerTabBarActions} from '~/ContentEditor/editorTabs/registerTabBarActions';
-import {registerSidePanelTabs} from '~/ContentEditor/editorTabs/EditPanelContent/SidePanel/registerSidePanelTabs';
+import {registerSidePanelTabs} from '~/JContent/SidePanel/registerSidePanelTabs';
 import {goBackAction} from './actions/contenteditor/goBackAction';
 
 export const registerActions = registry => {
