+++ src/javascript/JContent/ContentRoute/ContentLayout/ContentLayout.scss
@@ -15,6 +15,7 @@
     display: flex;
     flex: 1 1 0;
     min-width: 0;
+    overflow: hidden;
 
     background-color: var(--color-gray_light);
 }
