+++ src/javascript/JContent/SidePanel/ContentDetails/ContentDetails.scss
