+++ src/javascript/JContent/SidePanel/ContentDetails/index.js
