+++ src/javascript/JContent/SidePanel/ContentHistory/ContentHistory.scss
