+++ src/javascript/JContent/SidePanel/ContentHistory/index.js
