+++ src/javascript/JContent/SidePanel/SidePanel.jsx
