+++ src/javascript/JContent/SidePanel/SidePanel.scss
