+++ src/javascript/JContent/SidePanel/SidePanelContext.js
