+++ src/javascript/JContent/SidePanel/index.js
