+++ src/javascript/JContent/SidePanel/sidePanelTabAction.jsx
